+++ src/BorrowLogic/BorrowHandlers.sol
@@ -89,8 +89,7 @@ abstract contract BorrowHandlers is IBorrowHandlers, BorrowCheckers, SafeMint {
             nft: nft,
             loanId: ++proto.nbOfLoans // returns incremented value (also increments in storage)
         });
-        // todo #420 check args length not too long
-        for (uint8 i = 0; i < args.length; i++) {
+        for (uint256 i = 0; i < args.length; i++) {
             collatState = useOffer(args[i], collatState);
             lent += args[i].amount;
         }
@@ -108,7 +107,7 @@ abstract contract BorrowHandlers is IBorrowHandlers, BorrowCheckers, SafeMint {
             collateral: nft,
             supplyPositionIndex: supplyPositionIndex,
             payment: notPaid,
-            nbOfPositions: uint8(args.length)
+            nbOfPositions: args.length
         });
         proto.loan[collatState.loanId] = loan;
         emit Borrow(collatState.loanId, abi.encode(loan));
