+++ test/Repay.t.sol
@@ -18,13 +18,13 @@ contract TestRepay is External {
         repayNTimes(12);
     }
 
-    function repayNTimes(uint8 nbOfRepays) internal {
+    function repayNTimes(uint256 nbOfRepays) internal {
         uint256[] memory loanIds = new uint256[](nbOfRepays);
         uint256 balanceBorrowerBefore;
         uint256 toRepay = nbOfRepays * (uint256(1 ether * 2 weeks).mul(getTranche(0)) + 1 ether);
         uint256 currTokenId;
 
-        for (uint8 i = 0; i < nbOfRepays; i++) {
+        for (uint256 i = 0; i < nbOfRepays; i++) {
             loanIds[i] = i + 1;
             currTokenId = nft.mintOneTo(address(kairos));
             Loan memory loan = getLoan();
